+++ package.json
@@ -20,6 +20,7 @@
     "react-hook-form": "^7.51.2",
     "react-native": "0.72.4",
     "react-native-date-picker": "^5.0.0",
+    "react-native-image-picker": "^7.1.2",
     "react-native-maps": "^1.11.3",
     "react-native-maps-directions": "^1.9.0",
     "react-native-modal-selector": "^2.1.2",
